Fix resolving absolute paths
When calling the script from another folder, the script resolves absolute paths wrong, using its own location as a base, not the working directory, as expected. This fact could cause some nasty problems considering it's deleting stuff.

This is caused because it's resolving absolute paths prepending the `__dirname` value, but `path.join(__dirname, '.')` is not equivalent to the working directory in every case.

Node.js docs for the module scope:

    __dirname
    The directory name of the current module. This is the same as the path.dirname() of the __filename.

    __filename
    The file name of the current module. This is the current module file's absolute path with symlinks resolved.

    For a main program this is not necessarily the same as the file name used in the command line.

Node.js docs for the `fs` module:

    Relative paths will be resolved relative to the current working directory as determined by calling `process.cwd()`